+++ Sources/RVS_GeneralObserver/RVS_GeneralObserver_Protocols.swift
@@ -1,7 +1,7 @@
 /**
-    © Copyright 2019, The Great Rift Valley Software Company
+    © Copyright 2021, The Great Rift Valley Software Company
 
-    Verison: 1.0.4
+    Verison: 1.0.5
      
     LICENSE: